requirements.txt
ipython==8.22.2
ipdb==0.13.13
colorama==0.4.6
pytest==8.1.1
Jinja2==3.1.3
boto3==1.34.69
googlesearch-python==1.2.4

# AI and ML packages
tiktoken==0.6.0
openai==1.24.0
langchain == 0.1.17
langchain-community==0.0.36
langchain-openai==0.1.6
langchain-groq==0.1.3
langchain-aws>=0.1, <0.2.0
langfuse==2.29.0
